fixed bug that uploads file even though failed register
